requirements.txt
google-generativeai
fastapi
uvicorn
python-multipart
python-dotenv
librosa
numpy
scipy
soundfile
